Fix exp2(double) problem for SPIR-V
SPIR-V can handle a double-type input for exp2 with this change.

However, the slang-test is will failing to test it with an error message
saying, "abort compilation:".

With a simpler test case, I verified that SPIR-V assembly code is
properly generated for exp2(double) and I am not sure why slang-test is
still failing. We will need to revisit this issue later.

The simple testing is done with a following line:
  outputBuffer.result[0] = float(exp2(double(outputBuffer.result[0])));

And it generated following lines and it looks correct:
  ; Function main
  %main = OpFunction %void None %3
  %5 = OpLabel
  %16 = OpAccessChain %_ptr_Uniform_float %outputBuffer_0 %int_0 %uint_0
  %17 = OpLoad %float %16
  %19 = OpFConvert %double %17
  %20 = OpFConvert %float %19
  %21 = OpExtInst %float %1 Exp2 %20
  %22 = OpAccessChain %_ptr_Uniform_float %outputBuffer_0 %int_0 %uint_0
  OpStore %22 %21
  OpReturn
  OpFunctionEnd