Implement basic GLSL built-in functions (#3525)
* Implement basic GLSL built-in functions

Partially resolves #3362

This change implemented GLSL build-in functions described in the
following sections of "OpenGL Spec" document.
  8.1. Angle and Trigonometry Functions
  8.2. Exponential Functions
  8.3. Common Functions
  8.5. Geometric Functions
  8.7. Vector Relational Functions
  8.8. Integer Functions

About 40 functions are newly implemented and about 150 functions
were preexisted on HLSL side implementation.

The implementation of new functions hasn't been tested yet.

* Unify some of GLSL functions into hlsl.meta.slang

Partially resoves #3362

This change moves Some of GLSL functions from glsl.meta.slang to
hlsl.meta.slang, because those functions are generic enough to be used
for HLSL. Those functions are: dot, normalize, fma, and reflect.

There was "fma" for double in hlsl.meta.slang and it is converted to use
__BuiltinFloatingPointType type, which required some modifications in
diff.meta.slang. The implementation for "fma" in diff.meta.slang is very
similar to how "mad" is implemented.

* Implement more GLSL built-in functions

Partially resolves #3362

This change implements more GLSL built-in functions mentioned in the
following sections.
  8.4. Floating-Point Pack and Unpack Functions
  8.6. Matrix Functions

This change implemented 11 new GLSL built-in functions and there were 3
already working functions.

The mistake in "normalize" is fixed.

"refract" function is moved from glsl.meta.slang to hlsl.meta.slang.

* Implement basic GLSL built-in functions

Partially resolves #3362

This change implemented GLSL build-in functions described in the
following sections of "OpenGL Spec" document.
  8.1. Angle and Trigonometry Functions
  8.2. Exponential Functions
  8.3. Common Functions
  8.5. Geometric Functions
  8.7. Vector Relational Functions
  8.8. Integer Functions

About 40 functions are newly implemented and about 150 functions
were preexisted on HLSL side implementation.

The implementation of new functions hasn't been tested yet.

* Unify some of GLSL functions into hlsl.meta.slang

Partially resoves #3362

This change moves Some of GLSL functions from glsl.meta.slang to
hlsl.meta.slang, because those functions are generic enough to be used
for HLSL. Those functions are: dot, normalize, fma, and reflect.

There was "fma" for double in hlsl.meta.slang and it is converted to use
__BuiltinFloatingPointType type, which required some modifications in
diff.meta.slang. The implementation for "fma" in diff.meta.slang is very
similar to how "mad" is implemented.

* Implement more GLSL built-in functions

Partially resolves #3362

This change implements more GLSL built-in functions mentioned in the
following sections.
  8.4. Floating-Point Pack and Unpack Functions
  8.6. Matrix Functions

This change implemented 11 new GLSL built-in functions and there were 3
already working functions.

The mistake in "normalize" is fixed.

"refract" function is moved from glsl.meta.slang to hlsl.meta.slang.

* Fix a few minor bugs on GLSL builtin functions

Partially resovles #3362

Following bugs were addressed:
  1. "bitCounts" had to have a "Capability" on its function declaration.
  2. "roundEven" is implemented. It is almost same to "round()" but the
     behaivor is slightly different the given value is 1.5, 3.5, 5.5 and
     so on.
  3. umulExtended and imulExtended are simplified.
  4. exp2 is implemented with "__target_switch" for GLSL and SPIR-V.
  5. "tests/glsl-intrinsic/intrinsic-basic.slang" checks the results
     from the GLSL functions. Currently it is mainly to test if the
     functions exist or not, but it can now also test for a simple case
     where the input value is zero and the result is most of the time
     zero or one.

* Disable GLSL exp2 double type tests

This change disables some of GLSL exp2 related tests as a workaround.
The spir-v needs to handle the double-type argument for exp2 properly.

* Fix exp2(double) problem for SPIR-V

SPIR-V can handle a double-type input for exp2 with this change.

However, the slang-test is will failing to test it with an error message
saying, "abort compilation:".

With a simpler test case, I verified that SPIR-V assembly code is
properly generated for exp2(double) and I am not sure why slang-test is
still failing. We will need to revisit this issue later.

The simple testing is done with a following line:
  outputBuffer.result[0] = float(exp2(double(outputBuffer.result[0])));

And it generated following lines and it looks correct:
  ; Function main
  %main = OpFunction %void None %3
  %5 = OpLabel
  %16 = OpAccessChain %_ptr_Uniform_float %outputBuffer_0 %int_0 %uint_0
  %17 = OpLoad %float %16
  %19 = OpFConvert %double %17
  %20 = OpFConvert %float %19
  %21 = OpExtInst %float %1 Exp2 %20
  %22 = OpAccessChain %_ptr_Uniform_float %outputBuffer_0 %int_0 %uint_0
  OpStore %22 %21
  OpReturn
  OpFunctionEnd

* Add __floatCast that is safer than slang_noop_cast

Adding __floatCast that can be used for exp2 function.