Implement short-circuit logic operator
Implement short-circuit evaluation for logic && and ||
operator.

The short-circuit behavior is only used when the operands
involved are scalar and the parent function is non-differentiable.

In implementation, we define a new class 'LogicOperatorShortCircuitExpr'
derived from 'OperatorExpr'. In the visitInvoke() call, we will create
a new expression object 'LogicOperatorShortCircuitExpr' if the
expression is logic && or ||. So that we can generate new IR code in the
new visit function 'visitLogicOperatorShortCircuitExpr' to implement the
short-circuit behavior.

Add new test to test the short-circuit behavior.