Implement short-circuit logic operator (#3635)
* Implement short-circuit logic operator

Implement short-circuit evaluation for logic && and ||
operator.

The short-circuit behavior is only used when the operands
involved are scalar and the parent function is non-differentiable.

In implementation, we define a new class 'LogicOperatorShortCircuitExpr'
derived from 'OperatorExpr'. In the visitInvoke() call, we will create
a new expression object 'LogicOperatorShortCircuitExpr' if the
expression is logic && or ||. So that we can generate new IR code in the
new visit function 'visitLogicOperatorShortCircuitExpr' to implement the
short-circuit behavior.

Add new test to test the short-circuit behavior.

* Fix an compile issue occurred in Falcon test

Previously, we early return when at least one of the operands of
"&&" or "||" is vector in convertToLogicOperatorExpr call. However,
in that case the arguments involved in the expression have already been
type checked. When it falls-back to 'visitInvokeExpr', it will check
the arguments again, and some unexpected behavior could occur
which could in turn cause some internal error.

So we add a check in the 'visitInvokeExpr' to avoid double type checking
of arguments.

* Update glsl subgroup test to not use short-circuit

Since the short-circuit evaluation could cause the threads
diverging in subgroup intrinsics. So change the test to not
using "&&" to chain those subgroup intrinsics together. Instead,
using "&" to chain them together because those test functions have
the return value as bool.

* Disable short-circuit in few situations

Disable short-circuit in following situations:
1. generic parameter list
2. static const varible initialization

* Use a flag to indicate the enablement of short-circuit

Instead of using a struct to indicate the state of the outer
environment of current expression, use a simple bool flag to
indicate whether or not apply the short-circuit to current
expression because there few situations where we will disable
short-circuiting and in those circumstances, there is no nested.

Therefore, a flag is good enough to indicate the case.

* Disable short-circuit in index expression

Also fix the build issue. (A cleanup for the last change.)

* check both 'static' and 'const' modifiers

Previously we only check HLSLStaticModifier to decide whether or
not using short-circuit, but we really should check both 'static'
and 'const' modifiers together, because we only want to disable
the short circuit for init expression for 'static const' variable.

* relax the restriction of short-circuit for index expression

Disable the short-circuit for index expression only when declare
an array.

* Simplify the logic by creating subVisitor

Simplify the logic by create a sub expression visitor so
that we don't need to introduce extra recursion.

* Call convertToLogicOperatorExpr after args check

Change to call convertToLogicOperatorExpr after arguments
check in visitInvokeExpr such that we don't have to check
whether the arguments checked to avoid the double checking
issue.