Fix incorrect SPV stride for unsized array (#3837)
* Fix incorrect SPV stride for unsized array (#3825)

In '-emit-spirv-directly' mode, slang generates the stride 0
for unsized array in `OpDecorate` instructions.

For unsized array, the stride is invalid, but we need to provide
a non-zero value to pass the spirv validator.

* Decorate struct with unsized array field as 'Block'

For the struct having unsized array fields, it has to be decorated
as "Block", otherwise it will fails the spirv-val.

So we add a check at in 'emitGlobalInst' when emitting spirv for
'kIROp_StructType', where if there is unsized array field inside
the struct, emit a decorate instruction for above purpose.

* Update decoration for kIROp_SizeAndAlignmentDecoration

When add a decoration node for kIROp_SizeAndAlignmentDecoration,
we implicitly convert the 64 bit size to 32 bit. In most cases, this
should not be a problem because we won't have that large data type.
However, we use 64-bit -1 to represent the size of unsized-array,
so in that case, the conversion will change the size to 0, which is
incorrect. So change that decoration to use 64-bit size.

---------

Co-authored-by: Yong He <yonghe@outlook.com>