Implement if(let ...) syntax
Add "struct IfLetStmt: IfStmt" to define the
if "(let xxx = xxxxx)" syntax.

In parsing the predicate expression, we check if the identifier
is "let", then parse the "let" expression and return it to predicate.

In parseIfStatement(), we will check if the predicate expression is a
'let' expression. If so, we will create an IfLetStmt.