Change to not use IfLetStmt struct
Change to not use 'IfLetStmt', instead we will use a SeqStmt to store
DeclStmt and IfStmt, so that we don't need to change the semantics
checking and ir lowering.

In addition, in the positive statement, we insert another DeclStmt;

let var = OptVal.value

so user can use the var directly.

Therefore, the de-sugaring will transform the code into

if (let var = X as Y) =>

{
    let optVar: X as Y
    if (optVal.hasValue == true)
    {
        let var = optVal.value
    }
}