Updating CONTRIBUTION guide to use CMake
Releated to #3703

Removing the build instruction with Premake and replacing it with an
instruction with CMake. It is because we are going to move over to CMake
anytime soon.

Bumping the required CMake version to 3.25.0. When CMakePresets.json
has "version:6" it requires CMake version to be 3.25 or above. See the
URL below for more information,
  https://cmake.org/cmake/help/latest/release/3.25.html