Updating CONTRIBUTION guide to use CMake (#4017)
Releated to #3703

Removing the build instruction with Premake and replacing it with an
instruction with CMake. It is because we are going to move over to CMake
anytime soon.

Bumping the required CMake version to 3.25.0. When CMakePresets.json
has "version:6", it requires CMake version to be 3.25 or above. See the
URL below for more information,
  https://cmake.org/cmake/help/latest/release/3.25.html

CMakeLists.txt copies the prebuilt binary files from
  external/slang-binaries/bin/windows-x64

CMakeLists.txt was copying "slang-llvm.dll" to build/Release/lib
directory when it should have been build/Release/bin. It made slang-test
to ignore all FILECHECK tests. This is fixed.

Co-authored-by: Yong He <yonghe@outlook.com>