Keep const-ness in generic functions (#4028)
* Keep const-ness in generic functions

Closes #3834

The issue was that "const" variables inside of generic functions became
non-const variables.  This issue prevented some of GLSL texture
functions from being called inside of generic functions.

When `propagateConstExpr()` iterates the global functions, the generic
functions had to be handled little differently. This commit allows the
iteration to happen for the generic functions.

* Adding an explantion of the test as a comment