WIP: Force Inline If RefType (#4005)
* Force Inline if reftype

Fixes #3997.
If we are using a refType, we now ForceInline.

remarks:
1. Modifications were made in slang-ir-glsl-legalize to change how we translate GlobalParam proxy's into GlobalParam.
 a. We now handle the senario where a globalParam is used in multiple disjoint blocks (like 2 different functions).

* try to figure out why CI fails but local works

try to inline DispatchMesh, works locally, may fail on CI(?)

* try another fix

* add task tests + don't allow semi-early task-shader inline

Task shader uses DispatchMesh which is a very big 'hack' where we check for the function name and modify the callees in very large ways. This function does inline, but it cannot inline early due to future mangling that this operation requires todo. This is reflected with the `[noRefInline]` modifier. It is a modifier so users may stop mandatory inlines with `__ref` parameter.