Fix invoke resolution when dealing with overloded type expressions
This specific issue comes up with the `Differential` associated types where the concrete implementation and the associated type are both present in an overload group.

The fix is to simply resolve the overload before handling the type, since associated types are not proper type definitions.

Fixes https://github.com/shader-slang/slang/issues/3426