Do not mangle the name of identifiers when __extern_cpp is added (#4052)
Do not mange the name of identifiers decorated by "__extern_cpp".

For a slang files that are included by the library module and entry
point module, slang could generated two different mangled names for
the same functions, because the function with a struct parameter will
make the mangled function name contains the file name. Therefore, we
allow using "__extern_cpp" on such struct, such that no file name is
associated in the mangled name.