SPIRV: Fix storage class for unwrapped pointers
In SPIRV legalization, a struct wrapper is created around push constants
but is not itself legalized. Putting the struct type into the work
list causes the storage access of the push constant pointer to be
PhysicalStorageBuffer as expected, instead of Function scope that was
produced without the added struct legalization.

Adds a SPIRV test that exercises the fix.

Fixes #3946