Copy default target's optionSet to code-gen target's optionSet
In current implementation, the some options will be to added
to the target that is only specified by command line "-target".
But if user specifies the target by just using slang API,
e.g. 'spAddCodeGenTarget', those options will be missed.

To fix the problem, we copy the default target's options to
the code-gen target's option set. The default target will only be
useful when there is no target specified in the command line.