Fix race-condition and visual artifacts issues (#4152)
* Fix race-condition and visual artifacts issues

In PerformanceProfiler::getProfiler() we return a static object
for the profiler implementation, this is not thread-safe, so change
it to thead_local.

There is still some visual artifacts when using slang as the shading
language. We don't know the root cause yet, but found out it's related
to our loop inversion algorithm. So stage this feature for now, and turn
it into an internal option and default off. We will re-enable it after
more investigation on this optimization.

File an new issue 4151 to track it.

* Add '-loop-inversion' to the few tests