Support combined textures for Metal target
This commit is to support the combined texture cases where the sampler
is not given for the texture fetching. Internally it creates a default
sampler and use it for the fetching. The logic is mostly same to how it
works for HLSL.

It turned out that we were missing an offset variant of `Sample()`
function. This commit adds it.

This commit also fixes the cube texture related testing problems in
`tests/metal/texture.slang`.