Support int and uint typed samplers for GLSL texture functions
Closes #4196

The issue was introduced when we tried to support the sampler-less
textures for GLSL texture functions. And some of GLSL functions could no
longer take int/uint typed sampler anymore.

This commit fixes both problems.

One of the findings is that GLSL supports only three types of samplers:
`samplerXX`, `isamplerXX` and `usamplerXX`. They are not variable sized
vectors and it means we need to support only three types with fixed
number of components: vector<float,4>, vector<int,4> and vector<uint,4>.

Depth textures are always float type.