CTS: do not continue on error
The CTS nightly stops report error because of `continue-on-error` is set
to true. Set it false such that we can know as soon as it fails.

TODO:
Temporarily change the workflow to run per-PR to verify thing is
correct. Will change it back when the PR is ready to merge.