Support SPIR-V DebugTypePointer
Closes #4077

This commit is to emit DebugTypePointer when targeting spirv.

Currently DebugTypePointer can take a type decleared by DebugTypeBasic
and it doesn't work with a type declared by DebugTypeComposite. For this
reason, the function "isSimpleDataType" returns true for a memory
pointer type whose base type is the BasicType such as float, int, uint
and so on; it return false when the base type of the pointer is a struct
or vector.