Simplify the profiler and provide a 'clear' option
Change the profiler design to only return the thread_local
profiler to user.

We create a ISlangProfiler interface to carry the thread_local
variable PerformanceProfilerImpl profiler to user.

In addition, we provide a new option in the input parameter to
control whether or not user want to clear the previous profile
data. So spGetCompileProfile() can always returns a fresh new
profiling data.