Add APIs to get profile of compile time (#4242)
* Add APIs to get profile of compile time

Add serial time measurement

Add profiler to measure lots of stages in slang compilation, and it
can accumulate the time spent in each thread in multi-threads case and
finally report a serial timing info.

* Add invocation times to the profiler

* Simplify the profiler and provide a 'clear' option

Change the profiler design to only return the thread_local
profiler to user.

We create a ISlangProfiler interface to carry the thread_local
variable PerformanceProfilerImpl profiler to user.

In addition, we provide a new option in the input parameter to
control whether or not user want to clear the previous profile
data. So spGetCompileProfile() can always returns a fresh new
profiling data.

* Change to use slang container List

Stop using std::vector, instead use slang's container List.
Generate a UUID for ISlangProfiler